+++ src/MetaheuristicsPlatform/Algorithms/Multiobjective/Moeadd/MoeaddReferences.cs
@@ -1,0 +1,12 @@
+using MetaheuristicsPlatform.Classification;
+namespace MetaheuristicsPlatform.Algorithms.Multiobjective.Moeadd;
+public static class MoeaddReferences
+{
+    public static ScientificReference LiDebZhangKwong2015 { get; } =
+        new(
+            "Li, Deb, Zhang & Kwong",
+            2015,
+            "An Evolutionary Many-Objective Optimization Algorithm Based on Dominance and Decomposition",
+            "IEEE Transactions on Evolutionary Computation 19(5), 694-716",
+            "10.1109/TEVC.2014.2373386");
+}